Make sidebar scrollable and sticky (on modern browsers) (#91)
Just make the sidebar stick to the top with CSS and show a scrollbar on
overflow. There's no fallback for older browsers -- since up to now,
this didn't work for anyone (because "intelligent" scrolling has been
broken for years), I assume improving the UX at least for modern
browsers is still a net improvement.